requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.52
pydantic==1.10.7
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
alembic==1.11.1
pymysql==1.1.0
email-validator==1.3.1
